==> SPY_FBProphet_Project/Code/10_Conclusion.py <==
conclusion_text = """
Based on the results from both models, it's evident that a buy-and-hold strategy is a much better approach for long-term profitability compared to the predictions made by these models. However, the 80% model significantly outperformed the 50% model, which suggests potential in more refined models. This improvement piques my interest in conducting more extensive projects with FB Prophet, potentially incorporating additional factors, to see if a profitable predictive model can be achieved.
"""

print(conclusion_text)
